+++ app/controllers/Utils.scala
@@ -3,9 +3,10 @@ package controllers
 import java.net.URL
 import models._
 import play.api.data.format.Formatter
-import play.api.data.{Mapping, Forms, FormError}
-import play.api.mvc.{RequestHeader, Request}
+import play.api.data.{FormError, Forms, Mapping}
+import play.api.mvc.{Request, RequestHeader}
 import org.apache.commons.lang.StringEscapeUtils
+import play.Play
 
 import scala.collection.mutable.ListBuffer
 
@@ -14,9 +15,12 @@ object Utils {
    * Return base url given a request. This will add http or https to the front, for example
    * https://localhost:9443 will be returned if it is using https.
    */
-  def baseUrl(request: Request[Any], absolute: Boolean = true) = {
+  def baseUrl(request: RequestHeader, absolute: Boolean = true) = {
     if (absolute) {
-      routes.Files.list().absoluteURL(https(request))(request).replace("/files", "")
+      Play.application.configuration.getString("clowder.url") match {
+        case s:String => s
+        case _ => routes.Files.list().absoluteURL(https(request))(request).replace("/files", "")
+      }
     } else {
       routes.Files.list().url.replace("/files", "")
     }