+++ app/services/CrowdProvider.scala
@@ -31,7 +31,7 @@ class CrowdProvider(application: Application) extends IdentityProvider(applicati
       val token = UUID.generate().stringify
       val userip = request.remoteAddress
       UserService.save(new Token(token, userip, DateTime.now(), DateTime.now().plusMinutes(CrowdProvider.crowdServerTimeOut), false))
-      val query = Map[String, Seq[String]]("redirecturl" -> Seq[String](RoutesHelper.authenticate(id).absoluteURL()),
+      val query = Map[String, Seq[String]]("redirecturl" -> Seq[String](controllers.Utils.baseUrl(request) + RoutesHelper.authenticate(id)),
         "token" -> Seq[String](token))
       Left(Results.Redirect(CrowdProvider.crowdServerURL, query))
     } else {